requirements.txt
torch
numpy
streamlit
pillow
albumentations
torchvision